+++ sdk/go/core/tests/types_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/suite"
 	"strconv"
 	"testing"
+	"sort"
 )
 
 type SanityTypesTestSuite struct {
@@ -571,6 +572,181 @@ func (suite *SanityTypesTestSuite) TestListNoKeys() {
 		count = count + elem.Test.(string)
 	}
 	suite.Equal(count, "t1t2t3")
+}
+
+func checkEmptyStringValue(v string) string {
+	if len(v) == 0 {
+		v = "Exists"
+	}
+	return v
+}
+
+func printDictionary(legend string, entdict map[string]string) {
+	fmt.Printf("\n------> DICTIONARY%s\n", legend)
+	var keys []string
+	for k := range entdict {
+	        keys = append(keys, k)
+	}
+	sort.Strings(keys)
+	for _, k := range keys {
+		fmt.Printf("%s: %s\n", k, checkEmptyStringValue(entdict[k]))
+	}
+}
+
+func printDiff(diff map[string]types.StringPair) {
+	fmt.Printf("\n------> DIFFS:\n")
+	var keys []string
+	for k := range diff {
+	        keys = append(keys, k)
+	}
+	sort.Strings(keys)
+	for _, key := range keys {
+		valuePair := diff[key]
+		fmt.Printf("%s: %s vs %s\n", key,
+			checkEmptyStringValue(valuePair.Left),
+			checkEmptyStringValue(valuePair.Right))
+	}
+}
+
+func (suite *SanityTypesTestSuite) TestEntityDiff_TwoKeys() {
+	runner1 := ysanity.Runner{}
+	l1 := ysanity.Runner_TwoKeyList{First: "f1", Second: 11, Property: 82}
+	l2 := ysanity.Runner_TwoKeyList{First: "f2", Second: 22, Property: 83}
+	runner1.TwoKeyList = []*ysanity.Runner_TwoKeyList {&l1, &l2}
+	types.SetAllParents(&runner1)
+
+	entDict1 := types.EntityToDict(&runner1)
+	suite.Equal(len(entDict1), 4)
+	printDictionary("-ORIGINAL", entDict1)
+
+	runner2 := ysanity.Runner{}
+	ll1 := ysanity.Runner_TwoKeyList{First: "f1", Second: 11, Property: 82}
+	ll2 := ysanity.Runner_TwoKeyList{First: "f2", Second: 22, Property: 83}
+	runner2.TwoKeyList = []*ysanity.Runner_TwoKeyList {&ll1, &ll2}
+	types.SetAllParents(&runner2)
+
+	diff := types.EntityDiff(&runner1, &runner2)
+	suite.Equal(0, len(diff))
+
+	ll1.Property = 83
+	entDict2 := types.EntityToDict(&runner2)
+	printDictionary("-CHANGED-LEAF", entDict2)
+
+	diff = types.EntityDiff(&runner1, &runner2)
+	suite.Equal(1, len(diff))
+	printDiff(diff)
+
+	ll1.First = "f2"
+	entDict3 := types.EntityToDict(&runner2)
+	printDictionary("-CHANGED-KEY", entDict3)
+
+	diff = types.EntityDiff(&runner1, &runner2)
+	suite.Equal(2, len(diff))
+	printDiff(diff)
+}
+
+func (suite *SanityTypesTestSuite) TestEntityDiff_Presence() {
+	runner := ysanity.Runner{}
+	runner.Runner2.SomeLeaf = "some-leaf"
+	types.SetPresenceFlag(&runner.Runner2)
+	types.SetAllParents(&runner)
+
+	entDict1 := types.EntityToDict(&runner)
+	suite.Equal(len(entDict1), 2)
+	printDictionary("-LEFT", entDict1)
+
+	emptyRunner := ysanity.Runner{}
+	types.SetAllParents(&emptyRunner)
+	entDict2 := types.EntityToDict(&emptyRunner)
+	suite.Equal(len(entDict2), 0)
+	printDictionary("-RIGHT", entDict2);
+
+	diff := types.EntityDiff(&runner, &emptyRunner)
+	suite.Equal(1, len(diff))
+	printDiff(diff)
+}
+
+func (suite *SanityTypesTestSuite) TestEntityDiff_TwoListPos() {
+	runner1 := ysanity.Runner{}
+	elem1 := ysanity.Runner_TwoList_Ldata{}
+	elem2 := ysanity.Runner_TwoList_Ldata{}
+	elem1.Number = 1
+	elem1.Name = "runner:twolist:ldata:1"
+	elem2.Number = 2
+	elem2.Name = "runner:twolist:ldata:2"
+
+	elem11 := ysanity.Runner_TwoList_Ldata_Subl1{}
+	elem12 := ysanity.Runner_TwoList_Ldata_Subl1{}
+	elem11.Number = 11
+	elem11.Name = "runner:twolist:ldata:1:subl1:11"
+	elem12.Number = 12
+	elem12.Name = "runner:twolist:ldata:1:subl1:12"
+
+	elem1.Subl1 = append(elem1.Subl1, &elem11)
+	elem1.Subl1 = append(elem1.Subl1, &elem12)
+
+	elem21 := ysanity.Runner_TwoList_Ldata_Subl1{}
+	elem22 := ysanity.Runner_TwoList_Ldata_Subl1{}
+	elem21.Number = 21
+	elem21.Name = "runner:twolist:ldata:2:subl1:21"
+	elem22.Number = 22
+	elem22.Name = "runner:twolist:ldata:2:subl1:22"
+
+	elem2.Subl1 = append(elem2.Subl1, &elem21)
+	elem2.Subl1 = append(elem2.Subl1, &elem22)
+
+	runner1.TwoList.Ldata = append(runner1.TwoList.Ldata, &elem1)
+	runner1.TwoList.Ldata = append(runner1.TwoList.Ldata, &elem2)
+	types.SetAllParents(&runner1)
+
+	entDict := types.EntityToDict(&runner1)
+	suite.Equal(12, len(entDict))
+	printDictionary("-LEFT", entDict)
+
+	runner2 := ysanity.Runner{}
+	elem1 = ysanity.Runner_TwoList_Ldata{}
+	elem1.Number = 1
+	elem1.Name = "runner:twolist:ldata:1"
+	elem11 = ysanity.Runner_TwoList_Ldata_Subl1{}
+	elem12 = ysanity.Runner_TwoList_Ldata_Subl1{}
+	elem11.Number = 11
+	elem11.Name = "runner:twolist:ldata:1:subl1:11"
+	elem12.Number = 12
+	elem12.Name = "runner:twolist:ldata:1:subl1:12"
+
+	elem1.Subl1 = append(elem1.Subl1, &elem11)
+	elem1.Subl1 = append(elem1.Subl1, &elem12)
+
+	runner2.TwoList.Ldata = append(runner2.TwoList.Ldata, &elem1)
+	types.SetAllParents(&runner2)
+	
+	entDict = types.EntityToDict(&runner2)
+	suite.Equal(6, len(entDict))
+	printDictionary("-RIGHT", entDict)
+
+	diff := types.EntityDiff(&runner1, &runner2)
+	suite.Equal(1, len(diff))
+	printDiff(diff)
+}
+
+func (suite *SanityTypesTestSuite) TestEntityDiff_OneAugList() {
+	oneAugList := ysanity.Runner_OneList_OneAugList{}
+	oneAugList.Enabled = true
+
+	elem1 := ysanity.Runner_OneList_OneAugList_Ldata{}
+	elem1.Number = 1
+	elem1.Name = "elem1"
+
+	elem2 := ysanity.Runner_OneList_OneAugList_Ldata{}
+	elem2.Number = 2
+	elem2.Name = "elem2"
+
+	oneAugList.Ldata = []*ysanity.Runner_OneList_OneAugList_Ldata {&elem1, &elem2}
+	types.SetAllParents(&oneAugList)
+
+	entDict := types.EntityToDict(&oneAugList)
+	suite.Equal(5, len(entDict))
+	printDictionary("", entDict)
 }
 
 func testParams(testName string, entity types.Entity) types.Entity {